requirements/base.txt
bokeh==3.4.3
certifi==2025.4.26
charset-normalizer==3.4.2
contourpy==1.3.0
coreapi==2.3.3
coreschema==0.0.4
coverage==7.8.0
Django==2.2.28
django-autocomplete-light==3.9.4
django-cors-headers==3.11.0
django-debug-toolbar==3.2.4
djangorestframework==3.13.1
exceptiongroup==1.2.2
fatslim==0.2.2
future==1.0.0
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
itypes==1.2.0
Jinja2==3.1.6
Markdown==3.8
MarkupSafe==3.0.2
mysqlclient==2.2.7
netifaces==0.11.0
numpy==2.0.2
packaging==25.0
pandas==2.2.3
pillow==11.2.1
pluggy==1.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
simplejson==3.20.1
six==1.17.0
sqlparse==0.5.3
tomli==2.2.1
tornado==6.4.2
tzdata==2025.2
Unidecode==1.4.0
uritemplate==4.1.1
urllib3==2.4.0
xyzservices==2025.4.0
zipp==3.21.0
